chore: fix golangci.yml issues (timeout, format) (#23)
Found that the `deadline: 10m` was incorrect, the correct way to set the
timeout is `timeout: 10m` (per
https://golangci-lint.run/usage/configuration/). I think this should
address the issue that popped up in running the github actions.

While fixing this, also found that the way we set the format was
incorrect, `formats` requires an array, we were passing a string.

Also included changes from running `make` post bumping down
controller-gen versions while bumping down go.mod version to 1.21.