chore: moves authorino configuration to struct
Introduces PlatformAuthorizationConfiguration struct which is used while
creating Authorization controller. This data is mandatory for the
reconciliation to work so it should belong to a struct which holds this
responsiblity.

It's left up to the caller how the required configuration is obtained. In the
main.go boostrap logic it is fetched from environment variables which
are backed by a config map (and have defaults).