feat(envtest): introduces a facade for Kubernetes EnvTest
This change introduces a facade for configuring and using Kubernetes
Env Test [1] library used for integration testing in this project. The
main goal of this enhancement is to reduce boilerplate of test suite
bootstrap code by hiding the details of EnvTest setup and
wiring the controller-under-test.

For this purpose `k8senvtest` package is introduce with the simple means
of configuring integration test suite, so the bootstrap logic is limited
to:

```golang
envTest = k8senvtest.Configure(
		k8senvtest.WithCRDs(filepath.Join(test.ProjectRoot(), "config", "crd", "external")),
		k8senvtest.WithScheme(testScheme),
	).
		WithControllers(ctrlUnderTest.SetupWithManager).
		Start(ctx)
```

Additionally, the facade struct gives an unified way of manipulating
cluster under test as it is composed of k8s client.Client and custom
Cleaner (see its godoc for details why it is needed) struct.

[1] https://book.kubebuilder.io/reference/envtest