chore(test): simplifies test structure
By extracting common part describing requirements for a component
to become part of platform routing it becomes easier to understand
what are the requirements for a component to be part of platform
routing. These pieces were redundant and not easy to immediately graps
as there was quite some boilerplate code overshadowing the actual
narrative.