chore(ctrl): adds names to controllers (#37)
Sets a unique name for each controller per CR in the form of
`ctrlName-kind`, where kind is Kind of watched resource, i.e.
`routing-notebook`.

This name shows up in metrics, which can allow us to have more granular
data.

When `Named` is not called, the controller takes its name from the
resource it is intended for (i.e. the one defined in `For` builder
chain). Without prefixing we would end up with conflated metrics when
both authorization and routing controllers are enabled for the same
kind.

Applies a similar structure to ctrl's logger with its name and
additional `component:kind` value.