feat(routing): handles deletion of watched resources (#39)
## Summary

This PR adds functionality to the routing controller so that now when a
watched resource is deleted, the routing objects created by the
controller within the gateway NS are cleaned up. Since the watched
resource will be in a separate namespace from the gateway NS, this is
needed since ownerReferences are not suitable for the cross-namespace
scenario.

We accomplish this by adding a finalizer to the watched resource,
blocking deletion of the resource until the finalizer is removed by our
controller. This allows the controller to cleanup the created resources
associated with the watched object by searching in the gateway namespace
by the `owner-name`, `owner-kind` and `owner-uid` labels to ensure we
delete the correct resources before removing the finalizer from the
watched resource.

The PR also adds a test case for this, testing that the correct
resources are removed in the `public;external` case and also moves some
test fixtures out with the goal of enhancing readability of tests.

## Limitations

The current implementation does not handle a scenario where watch CR has
annotation value changed or it's been entirely removed. This will come
as a separate feature.

Fixes: https://issues.redhat.com/browse/RHOAIENG-10727

---------

Co-authored-by: bartoszmajsak <bartosz.majsak@gmail.com>