feat: add configurable serviceSelector instead of hardcoded labels (#57)
Allow a component to configure the label selectors used to select the
services
that should be exposable. This allows Platform Routing to operator
on an unchanged target.

ServiceSelectors also allow templating of the key and value in the
configuration
to extract data from the Target CR resource.

Example:

"routing.opendatahub.io/{{.kind}}": "{{.metadata.name}}", // >
"routing.opendatahub.io/Service": "MyService"

> [!CAUTION]
> the target CR is represented as a map[string]any with lowercase
> so in contrast to the Object representation {{ .Name }} it's required
to
> write the full object path as {{ .metadata.name }}

---------

Co-authored-by: Bartosz Majsak <bartosz.majsak@gmail.com>