feat(routing): support multi port services
The selected Service might have multiple ports to different
types of apis or protocols. In the previous version we only
exported the first port under the name ServiceName+ServiceNamespace.

With this change we include all ports on the service and export
them as ServiceName+ServicePortName+SerivceNamespace

The PublicServiceName is changed to inclue the ServicePortName which
means we now also loop over each ServicePort in the exportedService.

Also moved the Host creation/concat into functions of the
TemplateData object which means that they are constructed in one location
and reused between templates and metadata updates in the api.