feat(controller): pass configuration in as part of activation (#75)
When running in embedded mode the controller is only created once but it
rely on configuration from updatable sources.

When e.g. the DSCI is updated, this does not get reflected to the
runnning controller.

Now the Operator can pass the new Configuration per reconcile to the
Controller

The Activable interface has expanded to take a generic any type as an
argument in the Activate(T) method. The Controller should store and use
this configuration for it's next reconcile.

---------

Co-authored-by: bartoszmajsak <bartosz.majsak@gmail.com>