ci: run tests only for test job (#77)
Running default `make` targets results in running linter before the
actual tests are executed. As we have separate job for linter we should
focus on running tests instead so we can get feedback for both aspects
at the same time when PR is opened/updated. Currently both jobs fail on
linter so tests are never executed.

In order to run tests for test job, `make test` target is invoked
instead.